+++ qiskit_class_converter/converters/quantum_circuit_to_matrix.py
@@ -17,8 +17,10 @@ QuantumCircuit to Matrix Converter
 #  KIND, either express or implied.  See the License for the
 #  specific language governing permissions and limitations
 #  under the License.
+from typing import List
 
 import numpy as np
+from qiskit import QuantumCircuit
 from qiskit.visualization import array_to_latex
 
 from qiskit_class_converter.converters.base import BaseConverter
@@ -32,7 +34,11 @@ class QuantumCircuitToMatrixConverter(BaseConverter):
     def actual_convert_action(self):
         self.logger.debug("quantum circuit to matrix")
         matrix_list = {"gate": []}
-        dag = self.qiskit.converters.circuit_to_dag(self.input_value)
+        # type validate
+        if isinstance(self.input_value, (List, QuantumCircuit)):
+            dag = self.qiskit.converters.circuit_to_dag(self.input_value)
+        else:
+            raise TypeError("QuantumCircuit is required.")
         for layer in dag.layers():
             circuit = self.qiskit.converters.dag_to_circuit(layer['graph'])
             matrix_list["gate"].append(self.qiskit.quantum_info.Operator(circuit).to_matrix())